Rename Moire effect component and update references
